+++ src/web/components/Home/RotateApiKeyCard.scss
